Rename launcher to Seattle & update everything
