+++ README.md
@@ -32,11 +32,11 @@ For the time being, the library is published to GitHub Packages
 
 And from 0.80.0 and onwards, it uses the [Compose Bill of Materials](https://developer.android.com/jetpack/compose/setup#bom-version-mapping) instead.
 
-
-| `metro` version   | BoM version |
-|-------------------|-------------|
-| 0.80.0 to 0.101.0 | 22.11.00    |
-| 0.102.0 to latest | 23.01.00    |
+| `metro` version    | BoM version |
+|--------------------|-------------|
+| 0.80.0 to 0.101.0  | 22.11.00    |
+| 0.102.0 to 0.104.0 | 23.01.00    |
+| 0.105.0 to latest  | 23.03.00    |
 
 
 ## Project Structure
