+++ build.gradle
@@ -1,8 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     ext {
-        compose_bom_version = "2023.01.00"
-        accompanist_version = "0.28.0"
+        compose_bom_version = "2023.03.00"
+        accompanist_version = "0.30.0"
         kotlin_version = "1.8.10"
         kotlin_compiler_version = "1.4.3"
         datastore_version = "1.0.0"
