+++ launcher/src/main/AndroidManifest.xml
@@ -7,12 +7,12 @@
                      tools:ignore="QueryAllPackagesPermission"/>
 
     <application
-        android:name=".MetroLauncherApp"
+        android:name="com.louis993546.seattle.SeattleApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MetroLauncher">
+        android:theme="@style/Theme.Seattle">
 
         <activity
             android:name=".MainActivity"
