+++ launcher/src/main/java/com/louis993546/seattle/SeattleApp.kt
@@ -1,9 +1,9 @@
-package com.louis993546.metro.launcher
+package com.louis993546.seattle
 
 import android.app.Application
 import timber.log.Timber
 
-class MetroLauncherApp : Application() {
+class SeattleApp : Application() {
     override fun onCreate() {
         super.onCreate()
         Timber.plant(Timber.DebugTree())
